backend/requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain-google-genai
langchain-tavily
yfinance
python-dotenv